checked out working code from -mocking branch
