+++ projectApp/test/__mocks__/fs.ts
@@ -1,4 +1,5 @@
 // code taken directly from https://vitest.dev/guide/mocking.html#file-system
 
-const { fs } = require('memfs')
+import {fs} from "memfs";
+
 module.exports = fs
