--- projectApp/test/hello.test.ts
@@ -1,38 +1,0 @@
-// code taken directly from https://vitest.dev/guide/mocking.html#file-system
-
-import { beforeEach, expect, it, vi } from 'vitest'
-import { fs, vol } from 'memfs'
-import { readHello } from './hello.ts'
-
-// tell vitest to use fs mock from __mocks__ folder
-// this can be done in a setup file if fs should always be mocked
-vi.mock('node:fs')
-vi.mock('node:fs/promises')
-
-beforeEach(() => {
-  // reset the state of in-memory fs
-  vol.reset()
-})
-
-it('should return correct text', () => {
-  const path = '/hello-world.txt'
-  fs.writeFileSync(path, 'hello world')
-
-  const text = readHello(path)
-  expect(text).toBe('hello world')
-})
-
-it('can return a value multiple times', () => {
-  // you can use vol.fromJSON to define several files
-  vol.fromJSON(
-    {
-      './dir1/hw.txt': 'hello dir1',
-      './dir2/hw.txt': 'hello dir2',
-    },
-    // default cwd
-    '/tmp',
-  )
-
-  expect(readHello('/tmp/dir1/hw.txt')).toBe('hello dir1')
-  expect(readHello('/tmp/dir2/hw.txt')).toBe('hello dir2')
-})
